# Conflicts:
#	api/src/main/java/api/converter/CartConverter.java
#	api/src/main/java/api/service/CartService.java
#	api/src/main/java/api/service/ReservationService.java
#	db/src/main/java/db/repository/CartRepository.java

+++ api/src/main/java/api/model/request/ReservationRequest.java
@@ -6,7 +6,6 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Data
@@ -20,5 +19,5 @@ public class ReservationRequest {
     private Integer capacity;
     private LocalDate startDate;
     private LocalDate endDate;
-    private BigDecimal totalPrice;
+    private Long totalPrice;
 }
